Create Webhook_Log
Script to log data from Particle console with webhook. Please follow the directions on the last page of the Lab 2 manual for setting up a webhook. You will need to deploy your google sheet as 'web app' after pasting this code, and then copy the URL that is given for the webhook setup.